Add tap gesture to provide additional feature info
I enhanced RoadmapFeature by adding a URL property that can be used to
display additional information about the feature request, such as a
GitHub issue or a blog post.

I enhanced RoadmapView by adding a bindable variable for the selected
feature in the list. I added a tap gesture recognizer that will update
the bindable value with the selected feature. This will allow a
developer to navigate to another view to show more detailed information
about the feature, or to use the new url field to open a web view to
show the detailed information about the feature.